Move WizConfigEntry to coordinator.py
