+++ homeassistant/components/wiz/binary_sensor.py
@@ -16,9 +16,8 @@ from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
-from . import WizConfigEntry
 from .const import DOMAIN, SIGNAL_WIZ_PIR
-from .coordinator import WizData
+from .coordinator import WizConfigEntry, WizData
 from .entity import WizEntity
 
 OCCUPANCY_UNIQUE_ID = "{}_occupancy"
