+++ homeassistant/components/wiz/fan.py
@@ -21,8 +21,7 @@ from homeassistant.util.percentage import (
     ranged_value_to_percentage,
 )
 
-from . import WizConfigEntry
-from .coordinator import WizData
+from .coordinator import WizConfigEntry, WizData
 from .entity import WizEntity
 
 PRESET_MODE_BREEZE = "breeze"
