+++ homeassistant/components/wiz/light.py
@@ -22,8 +22,7 @@ from homeassistant.components.light import (
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
-from . import WizConfigEntry
-from .coordinator import WizData
+from .coordinator import WizConfigEntry, WizData
 from .entity import WizToggleEntity
 
 RGB_WHITE_CHANNELS_COLOR_MODE = {1: ColorMode.RGBW, 2: ColorMode.RGBWW}
