+++ homeassistant/components/wiz/number.py
@@ -17,8 +17,7 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
-from . import WizConfigEntry
-from .coordinator import WizData
+from .coordinator import WizConfigEntry, WizData
 from .entity import WizEntity
