+++ homeassistant/components/wiz/sensor.py
@@ -16,8 +16,7 @@ from homeassistant.const import (
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
-from . import WizConfigEntry
-from .coordinator import WizData
+from .coordinator import WizConfigEntry, WizData
 from .entity import WizEntity
 
 SENSORS: tuple[SensorEntityDescription, ...] = (
